+++ Tests/ProjectDescriptionTests/Template/TemplateTests.swift
@@ -1,0 +1,28 @@
+import ProjectDescription
+import TuistSupportTesting
+import XCTest
+
+class TemplateTests: XCTestCase {
+    func test_template_codable() throws {
+        // Given
+        let template = Template(
+            description: "",
+            attributes: [
+                .required("name"),
+                .optional("aName", default: "defaultName"),
+                .optional("bName", default: ""),
+            ],
+            files: [
+                .string(path: "static.swift", contents: "content"),
+                .file(path: "generated.swift", templatePath: "generate.swift"),
+            ],
+            directories: [
+                "{{ name }}",
+                "directory",
+            ]
+        )
+
+        // Then
+        XCTAssertCodable(template)
+    }
+}